refactor(layout): streamline layout structure and remove unused theme provider, fixed conflict
